// src/generated/models/Crf82_slotsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf82_slotscrf82_validationstatus = {
  0: 'TRUE',
  1: 'FALSE'
} as const;
export type Crf82_slotscrf82_validationstatus = keyof typeof Crf82_slotscrf82_validationstatus;
export const Crf82_slotsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf82_slotsstatecode = keyof typeof Crf82_slotsstatecode;
export const Crf82_slotsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf82_slotsstatuscode = keyof typeof Crf82_slotsstatuscode;

export interface Crf82_slotsBase {
  crf82_additionalassistants?: string;
  crf82_campuslocation?: string;
  crf82_cancellationreason?: string;
  crf82_cancelledby?: string;
  crf82_coursename?: string;
  crf82_coursetype?: string;
  crf82_credithours?: number;
  crf82_endtime?: string;
  crf82_fulltimeindicator?: string;
  crf82_gradelevel?: string;
  crf82_instructorassistant?: string;
  crf82_iterationnumber?: number;
  crf82_legacyidentifier?: string;
  crf82_roomnumber?: string;
  crf82_sessiondate?: string;
  crf82_sessionstatus?: string;
  crf82_sessiontitle?: string;
  crf82_slotid: string;
  crf82_starttime?: string;
  crf82_stepcode?: string;
  crf82_validationstatus?: Crf82_slotscrf82_validationstatus;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf82_slotsstatecode;
  statuscode?: Crf82_slotsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf82_slots extends Crf82_slotsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf82_validationstatusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}